mariadb: change install paths from mariadb to mysql
Mariadb is advertised as a drop-in replacement for mysql and it
should behave like one. Some packages look for /usr/include/mysql and
/usr/${LIBDIR}/mysql. Since the packages already block each other
because of binary name clashes, collisions won't be a problem. So
it's either fixing above mentioned packages or just install to the
mysql paths.

Btw, this is also what Percona::CleverCloud and the big distros do.

Change-Id: Icae5f01aee138367ddb46272b698862e1fe03533