virtualbox-ose: use kBuild from system
virtualbox comes with a bundled kBuild version, which is used
if the directory exists. However, it seems this version has
issues when the build is using a newer gcc version.
Delete the directory in src_prepare to use the system-installed
kBuild, and update the dependency to require at least 0.1.999 as
specified by Config.kmk

Change-Id: Ie83382214a57f34786a81993c760a88e08021746
Reviewed-on: https://galileo.mailstation.de/gerrit/4070
Reviewed-by: Pierre Lejeune <superheron@gmail.com>